requirements.txt
langchain>=0.2.0
langchain-community
langchain-openai
langchain-cohere
chromadb
python-dotenv
fastmcp
pydantic
unstructured[md,docx,pdf,xlsx]
python-docx
pypdf
openpyxl
rank_bm25
click
tqdm
pytesseract
pdf2image
fastapi
uvicorn[standard]
mcp-client